Move IconButton into a package
