+++ packages/icon-button/src/availableTypes.js
@@ -1,0 +1,2 @@
+import types from './types';
+export default Object.values(types);
