+++ packages/icon-button/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './IconButton';
+export { default as types } from './types';
+export { default as AVAILABLE_TYPES } from './availableTypes';
