+++ packages/icon-button/src/types.js
@@ -1,0 +1,5 @@
+export default Object.freeze({
+  PRIMARY: "primary",
+  FLAT: "flat",
+  TRANSPARENT: "transparent"
+});
